feat: Refactor Settings and Attachments management
- Introduced a new Tabs component in the Settings page to organize account, API keys, and attachments sections.
- Removed the AttachmentsPage route and integrated its functionality directly into the Settings component.
- Updated routing to eliminate the standalone Attachments page, enhancing user navigation.
- Added attachment handling features, including upload, download, and deletion capabilities within the Settings interface.
- Updated package.json to include the latest version of "@tavily/core" and adjusted dependencies accordingly.
