+++ frontend/constants/routes.ts
@@ -3,7 +3,6 @@ export const ROUTES = {
   CHAT: "/chat",
   CHAT_THREAD: (threadId: string) => `/chat/${threadId}`,
   SETTINGS: "/settings",
-  ATTACHMENTS: "/attachments",
 } as const
 
 // Type helper for route paths
