+++ package.json
@@ -40,12 +40,13 @@
     "@radix-ui/react-slider": "1.2.2",
     "@radix-ui/react-slot": "latest",
     "@radix-ui/react-switch": "1.1.2",
-    "@radix-ui/react-tabs": "1.1.2",
+    "@radix-ui/react-tabs": "^1.1.12",
     "@radix-ui/react-toast": "1.2.4",
     "@radix-ui/react-toggle": "1.1.1",
     "@radix-ui/react-toggle-group": "1.1.1",
     "@radix-ui/react-tooltip": "latest",
     "@tailwindcss/typography": "latest",
+    "@tavily/core": "latest",
     "ai": "latest",
     "autoprefixer": "^10.4.20",
     "class-variance-authority": "^0.7.1",
@@ -81,7 +82,6 @@
     "tailwind-merge": "^2.5.5",
     "tailwind-scrollbar": "latest",
     "tailwindcss-animate": "^1.0.7",
-    "@tavily/core": "latest",
     "use-sync-external-store": "latest",
     "uuid": "latest",
     "vaul": "^0.9.6",
